quick fix (not really) for issue #43 https://github.com/cs150bf/ever-notedown/issues/43
Basically it's all because the css encapsulation business is very poorly
done in this package so there are bound to be all sorts of problems.
It's really a big mess. But anyway this should fix the "forced linebreaks for inline comment" problem
(for python code blocks) for now.

Note for self and other possible interested people (unrelated to this
issue #43):
- Checkout [markdown-it](https://github.com/markdown-it/markdown-it) and it math plugin [markdown-it-math](https://github.com/runarberg/markdown-it-math)